--- jupyter-config/nbconfig/notebook.d/data-grid.json
@@ -1,5 +1,0 @@
-{
-  "load_extensions": {
-    "data-grid/main": true
-  }
-}
